Fix build process for windows. Missing syscall object
